"YOUR GETTING HACKED" Pop-up
This is a pop-up that will activate when going to google classroom. This works on any computer that has the Tampermonkey extension. This only works on google classroom, but you can edit the URL next to // @match for it on any website that you put in. If you would like I could make more like this or add on to this one.